+++ CHANGELOG.md
@@ -12,6 +12,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 
+## [1.1.0] – 2025-06-23
+
+### Added
+
+- First WordPress REST API integration
+- Added cabins API fetch module `fetchCabins();`
+- Connected frontend to headless WordPress backend
+
+---
+
 ## [0.0.5] – 2025-06-22
 
 ### Added